final check of can vs had reconstruction and prep files
